Move seed to base class
